+++ packages/react-server-dom-webpack/src/__tests__/ReactFlightDOMReply-test.js
@@ -23,6 +23,10 @@ let ReactServerDOMClient;
 describe('ReactFlightDOMReply', () => {
   beforeEach(() => {
     jest.resetModules();
+    // Simulate the condition resolution
+    jest.mock('react-server-dom-webpack/server', () =>
+      require('react-server-dom-webpack/server.browser'),
+    );
     const WebpackMock = require('./utils/WebpackMock');
     // serverExports = WebpackMock.serverExports;
     webpackServerMap = WebpackMock.webpackServerMap;
